LayoutTests/fast/canvas/philip/tests/2d.text.font.parse.invalid.html

<!DOCTYPE html>
<title>Canvas test: 2d.text.font.parse.invalid</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<link rel="prev" href="2d.text.font.parse.system.html" title="2d.text.font.parse.system">
<link rel="next" href="2d.text.font.default.html" title="2d.text.font.default">
<body class="show_output">
<p>
 <a href="2d.text.font.parse.system.html" accesskey="p" title="[p] 2d.text.font.parse.system">&lt;</a>
 <a href="index.html">[index]</a>
 <a href="2d.text.font.default.html" accesskey="n" title="[n] 2d.text.font.default">&gt;</a>
<h1><a href="index.2d.html">2d</a>.<a href="index.2d.text.html">text</a>.<a href="index.2d.text.font.html">font</a>.<a href="index.2d.text.font.parse.html">parse</a>.invalid</h1>
<p class="desc"></p>
<div class="refs">References:
<ul>
<li><a href="spec.html#testrefs.2d.text.font.invalid">2d.text.font.invalid</a>

</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.font = '20px serif';
_assertSame(ctx.font, '20px serif', "ctx.font", "'20px serif'");

ctx.font = '20px serif';
ctx.font = 'bogus';
_assertSame(ctx.font, '20px serif', "ctx.font", "'20px serif'");

ctx.font = '20px serif';
ctx.font = 'inherit';
_assertSame(ctx.font, '20px serif', "ctx.font", "'20px serif'");

ctx.font = '20px serif';
ctx.font = '10px {bogus}';
_assertSame(ctx.font, '20px serif', "ctx.font", "'20px serif'");

ctx.font = '20px serif';
ctx.font = '10px initial';
_assertSame(ctx.font, '20px serif', "ctx.font", "'20px serif'");

ctx.font = '20px serif';
ctx.font = '10px default';
_assertSame(ctx.font, '20px serif', "ctx.font", "'20px serif'");

ctx.font = '20px serif';
ctx.font = '10px inherit';
_assertSame(ctx.font, '20px serif', "ctx.font", "'20px serif'");


});
</script>
