LayoutTests/platform/mac-snowleopard/canvas/philip/tests/2d.text.font.parse.invalid-expected.txt

< [index] >
2d.text.font.parse.invalid
References: 2d.text.font.invalid
Actual output:
Failed assertion ctx.font === '20px serif' (got inherit[string], expected 20px serif[string])
Failed assertion ctx.font === '20px serif' (got 10px initial[string], expected 20px serif[string])
Failed assertion ctx.font === '20px serif' (got 10px default[string], expected 20px serif[string])
Failed assertion ctx.font === '20px serif' (got 10px inherit[string], expected 20px serif[string])
